docs: Removed reference to draft apps.
When appropriate, changed references to "draft" apps to alpha and beta apps.
Left the "draft" info in the deprecated file v2/billing_integrate.jd, but
added a note that draft apps are no longer supported.

Docs are staged to:

in-app billing:
http://asolovay.mtv:9011/google/play/billing/billing_admin.html
http://asolovay.mtv:9011/google/play/billing/billing_testing.html

http://asolovay.mtv:9011/google/play/billing/v2/billing_integrate.html
(deprecated doc; just noted that draft is no longer supported)

LVL:
http://asolovay.mtv:9011/google/play/licensing/overview.html
http://asolovay.mtv:9011/google/play/licensing/licensing-reference.html

OBBs
http://asolovay.mtv:9011/google/play/expansion-files.html

As per a request from Dirk, I added a section to the "Testing In-App
Billing" page noting that draft apps are no longer supported, and
linking to the section on alpha and beta channels; I put in a link
to the new section in each place where we'd previously talked about
draft apps.

Also fixed formatting in a few code samples, where over-long lines were
forcing the code box to use a scroll bar.

bug: 15013770
Change-Id: I1368b9b5ebf2a0c22f84fcbcd5f37401b390c423

Former-commit-id: aee20f92b8e8fdc15bb7c1b01d831d55452398a0