Rebuild the enabled IMEs list when changed in Settings page
With this CL, the input method switching controller is notified
even when the list of enabled IMEs is changed by System
Settings.

Ideally we should hook every setter method in
InputMethodSettings to ensure that InputMethodSettings and
InputMethodSubtypeSwitchingController can share the same list.
Until then, most of user visible weirdness should disappear
with this CL.

BUG: 15299968
Change-Id: Ie41da226d71b58d2e12057f85e4793b35e8b9f23

Former-commit-id: 3d46bab02978df14354379c4b10d7f68c3b80771