Use more user actions to update the IME rotation order
Previously, only the text update event in composing mode
is used to update the IME rotation order. This may look strange
to a user especially when the character is input without
composition.

With this CL, to cover above cased, sending a key event is also
used as a trigger to update update the rotation order.

Note that the cost of calling IMM#notifyUserAction multiple times
is now cheap enough because of the optimization made in
I19ad8542659bc092b92ee13e.

BUG: 7043015
Change-Id: I03fa436df0a5e348b3f93170aab3a8ac5a7e1677

Former-commit-id: 39dd1510ea2bfb79709075f352384730862fd4c4