Add methods to TelephonyManager for Subscription management (1/3)
As part of the effort to permit third-party suppliers of telephone
connectivity, we permit clients of TelephonyManager to obtain the
currently valid Subscription objects describing these suppliers. The
Subscriptions are intended to be self-describing such that an end-user
can choose from among them based on the information they contain. One
of these objects may then be attached to an ACTION_DIAL or ACTION_CALL
Intent to represent the end-user's choice of how to make a call.

Change-Id: I2ad1930079042d7bfe762b47c22f0538f7403bff

Conflicts:
	telephony/java/com/android/internal/telephony/ITelephony.aidl

Conflicts:
	telecomm/java/com/android/internal/telecomm/ITelecommService.aidl

Former-commit-id: c35ad0252d1ad9ae50e2dadd670783d4a6259df6