TvContract: Add columns for video format and resolution
Added COLUMN_VIDEO_FORMAT to the Channels table to indicate typical
video format for programs from a given TV channel, which can later be
used to filter out channels. Also defined common video formats and
resolutions and a utility method to get the video resolution for a given
format.

Added COLUMN_VIDEO_WIDTH and COLUMN_VIDEO_HEIGHT to the Programs table
to indicate the actual video dimension for each individual TV program,
which can be shown as a part of program information by applications.

Bug: 15351111, 15456031
Change-Id: Id72e25de40f06a1b615620365e06cf11e4cd9f81

Former-commit-id: 68e0de81d5689d8f2b3c36cb067a43765557daa1