Remove KeySetManagerService lock.
The KeySetManagerService is a part of the PackageManagerService, which is
responsible for maintaining the association between packages, keysets and
public keys.  This information is a critical component of the
PackageManagerService data, and should require the holding of the
PackageManagerService's lock.  Remove the existing KeySetManagerService lock
and require the mPackages lock to be held instead.

Bug: 6967056
Change-Id: I803f8388e42469d30562b40212cf497320851268

Former-commit-id: d79fdf1c62671abac376c71aca8fa19484a1ecf2