docs: Added table for the L device images.
Table has current checksum images, but still needs download links.
I've tried out the "hammerhead" system image and it seems to work.

Also added ID names to the headings and fixed a couple of typos. As
per RobertLy@ 's request, took out the "how to flash" instructions
and just sent folks to the devsite page for instructions.

Doc staged to:

http://asolovay.mtv:9099/preview/setup-sdk.html#installImage

Change-Id: Ie60712c549de3c5416190a72de33619874dc6293

Former-commit-id: 0341fba92a8c2c73a6940834946528b74d0c4f9e