Text size for empty view shall be styled according to theme
In activity_list.xml the empty TextView has set a
textAppearance which allows the theme to set the font size.
However, for preference_list_fragment.xml the textAppearance
is missing but should follow the same style as
activity_list.xml

The missing textAppearance makes the empty text in
Settings->Wifi and Settings->Bluetooth look small if
no networks are found.

Change-Id: I9320b37cfeec0d9505d73650696eaf534fc4a321

Former-commit-id: 6526bc7634714def49fec067196ed552b8bc4933