Fix LocalePicker locale filtering.
This was broken by commit 407b595a4d1090351 [formerly b27c137087ee92674] which
removed a badly phrased check. Prior to that commit,
there was a check that excluded locales whose toString()
length was not 5.

This change reinstates that check in a less roundabout
way, by excluding BCP-47 tags that don't have a country
subtag.

bug: 16038949

Change-Id: Ic24b1df87b2e2ce47d9ae04c7759088721f27b4f

Former-commit-id: 80b574199c5d06132acbcf640ec81f45f0d76ce3