DO NOT MERGE Fix extraneous allocation and copying
With breaks being allocated the way it was, there were 16 ints with
value 0 being stored in the beginning of the vector. Because of the way
the rest of the code is structured, this did not result in incorrect
operation, but it still wasted time and memory.

Change-Id: Ic0df3e5484417da51f2465ec2d72222fefffc18a
(cherry picked from commit 733b3c0f8b77fefcd234c94170faa4e7bd76074c)

Former-commit-id: 1a408a344c7921d7c943826bea3ce0cdd822583d