Implement TvContentRating
Added a class that represents TV content ratings that are used for
parental controls. TV input services are required to create
TvContentRating objects with rating/sub-rating constants for specific
countries (for now US and KR only) defined in this class and then
convert them into strings in order to store them in the EPG database.

Bug: 13172379
Change-Id: I6200cb12fbc274dd0a8bcb87a8ac3cd2d96791db

Former-commit-id: e16873467cda89d905027458e2664c759465d463