Allow adding widgets from user profiles.
The goal of this change is to enable support for appwidget from
user profiles to the user main profile. A user profile is a user
which is associated as a child of the main user profile. For example,
a user may have a personal (parent) and corporate (child) profile.
The device policy should be able to control whether adding a widget
from a child profile and given packages is allowed. This change
assumes that all packages from managed profiles are white listed.
Another change will add the device policy changes.

(cherrypicked 8cd27c3c to work around Gerrit issue)

Change-Id: I267260b55d74c48b112a29979a9f59eef7a8194e

Former-commit-id: 39999cf0263d0568337a87d0e586dcafd9e22ba1