Restrict SCORE_NETWORKS permission to signature|system.
As long as the APIs aren't ready to be un-hidden, there's no need to
allow non-system apps to access them. This can be reverted when the
APIs are finalized and we have mechanisms in place for selecting
between scorers installed on the device, rather than only a
preselected scorer bundled with the system.

Change-Id: I745b4c691eccfb526014d4cb71639db2119e4fdb

Former-commit-id: 4bd7efbe982d26a68950f3c92bd0e7142c4ecbf7