Adding the ability for carrier app to override SPN and carrier name.
The change allows system apps (holding MODIFY_PHONE_STATE permission) or
carrier apps to override the "carrier branding" on a per ICCID basis. The
override affects the service provider name as well as the network operator
name. The override is also saved as a SharedPreference and will persist for
the iccId across reboots.

Change-Id: I985ba247e10e2501e3d0d21567ccadc46f365879

Former-commit-id: 86578cca383562c72751d747905280afdeb2bdfb