Add EXTRA_REPLACEMENT_EXTRAS for ACTION_CHOOSER
This allows the calling app to supply a set of alternative extras to
be sent along with the target intent if the user chooses specific
items from the full set available on the device. When replacement
extras have the same key as extras in the initial intent, the
replacement extras dominate.

Change-Id: I5d64c80447386f22402b71291bb289a74015d619

Former-commit-id: e49d9393783f2789239f8da536abda9533ffadf0