Remove COLUMN_CONDITIONAL_ACCESS from TvContract.
It turned out that having this flag in Channels table does not
make sense for some European countries because they scremble the signal
per program basis. Also, the current flag does not fit to the use-case
like 'pay per view'. Decided to remove this in L release because it's not
clear how we will handle the cases raised at this moment.

Change-Id: Ibbf829f331a307ac89a553e00ece18d5958e72b5

Former-commit-id: 94a509b123d1a115756bab8facf5d0ca54b42ff5