Crunch Roboto and fix its digits to have the same width.
Update to upstream version 2.001045, which adds fontcrunch to the font
production process and fixes a problem with uneven digit widths. Should fix
https://code.google.com/p/android-developer-preview/issues/detail?id=330.

The font footprint is reduced by 705KB, most of it due to the crunching, but
also by using a newer version of fonttools that removes extra unnecessary
paddings from the glyf table.

Bug: 15348046
Bug: 16296048
Bug: 16299966

Change-Id: I838711916f8c47db011cfe889b232e60e6f1aa0c

Former-commit-id: d847ebb8d736ce27a4abc59556ace8b29dcd90b3