Introduce startIsolatedProcess private API in ActivityManager
The new API spawns a isolated process, using a custom uid, entrypoint and
abi. Such API is used by the WebViewFactory to spawn its unpriviledged
but trusted process (hence the fixed uid) which rewrites the rerlo file
on boot / when an update occurs.
Since both the ActivityManager service and the WebViewUpdate service
live in the SystemServer their calls be dispatched locally and no
binder interface needs to be exposed for the new startIsolatedProcess API.

BUG:16403706
Change-Id: I327b59735c12698595e0dbcc4da5d759c9103b0a

Former-commit-id: f080c3a529a317258792a7c888abe3acf32601f2