Remove PendingActivityLaunches based on stack.
Was previously removing by trying to match activities, but the
activity being matched would not have been in the stack yet. Since
it doesn't get added to the stack when it goes into the list
of pending activities. By removing all pending activities associated
with the stack we clean up all such pending activities.

Fixes bug 16045752.

Change-Id: I6b5981ffcce674139837b06362e573299bbc9e52

Former-commit-id: 7f13ed37fd2a0533fa7c5e330519decd89fdc1d5