Prevent system server dump stuck by pipe buffer full.
Symptom:
Watchdog timeout during dumping db info.

Root Cause:
When the caller is system server, this invocation is not IPC
that means it is not asynchronous, it will need writer has
written all data to finish. But pipe has 16*4KB buffer limit,
if no reader consumes the data, it will keep waiting.

Solution:
Check if caller is system server, then use another thread to write
so reader could consume the data in buffer.

Change-Id: I4bf80fd645cc9396f51ffc0eb27fb895756c1dcf

Former-commit-id: 1d6c40a6698d271517e5b333f2d0dd4459913884