Give disappearing children a chance to recreate DisplayLists
Views which are on the secondary "disappearingChildren" list in a
ViewGroup can be displayed and animated, but are unable to properly
update their underlying DisplayLists. This change processes these
children the same as other children in a ViewGroup, enabling changes
in those children to be reflected in their DisplayLists.

Issue #11551606 disappearing views don't invalidate/redraw correctly

Change-Id: I8d016ada7b73f6a5e74c2f21c979c0b6f84703b1

Former-commit-id: 6c0665fd0f7717dc3301b0612860a5007d4ba2f3