CEC: Make deviceSelect to active device a no-op
Changing input to already the active source should return
immediately without initiating routing control. This is also required
for processing the incoming <Active Source>

Also set the flag in intent for input change to start a new
activity from a service. The service fails to launch an activity
without it.

Bug: 15570939
Change-Id: I9b20b31137dfa4dc847bc43cc7fd35e669d0dec7
Former-commit-id: 58500f43ecbed3f92d7c077fb6ce396252cd00ea