Properly update accessibility manager state.
Applciations that fire accessibility events have to first check
it accessibility is on and then fire the event. If the app fires
an event when the feature is off an exception is thrown. However,
due to the way accessibility state is pushed to the local
accessibility manager it was possible that the app checks that
aaccessibility is off, fires an event, then the state of the
local manager changes by the time the event is propagated up
the view tree and an exception is thrown. This is a regression.

We really want this exception to prevent apps sending events
across processes if accessibility is off, so now the state
we get from the system manager service is applied via a
message so if the app checks that accessibility is off, then
it will remain off (locally) until the event is propagated.

bug:16672964

Change-Id: Icf5d1b1bdff87b58f285f0d9e1a098552513bbe8

Former-commit-id: c946f14a67f36d398f30040976ec6e7baf2643ac