Switch CaptivePortalLogin app to use theme like Settings UI.
The theme is switched to Theme.Material.Settings.  The progress bar
window feature is not supported in Material (b/16652978) so I added
a progress bar to the layout.  The Theme.Material.Setting's accent
color is set such that ProgressBars are indistinguishable, so accent
color is reset back to the parent's setting.

bug:15409354
Change-Id: Ic2862b8439be8591ec426f3d4dffad72179b2539

Former-commit-id: 8f333f19222ac9415152e31f10e0df2b571b0b77