"Implement" formerly abstract methods VideoCallProvider.
Provide null-op methods so there is greater flexibility for classes
which extend this base class, but might not support all its methods.
In the future, if we add more methods to the *.aidl interface, this
class can work with binaries which extend this class but don't
implement all the methods since the superclass now implements those
methods.

Left onSetPreviewSurface and onSetDisplaySurface as abstract.

Bug: 16842159
Change-Id: Ie7fea6ef5d473e3924ef370392f380603a568800

Former-commit-id: 1f6bbd292050973aa6e2ac5a1669cd04822e06d1